용어,카테고리,답변 Entity 수정

+++ vita/src/main/java/com/ict/vita/service/termcategory/TermCategoryDto.java
@@ -20,8 +20,8 @@ public class TermCategoryDto {
 	private TermsDto termsDto; //용어
 	private String category; //카테고리명
 	private String description; //용어에 대한 설명
-	private int count = 0; //해당 용어에 속하는 데이터 갯수
-	private int parent = 0; //부모 용어id(0이면 최상위 부모다)
+	private long count = 0; //해당 용어에 속하는 데이터 갯수
+	private long parent = 0; //부모 용어id(0이면 최상위 부모다)
 	
 	//[TermCategoryDto를 TermCategoryEntity로 변환하는 메서드]
 	public TermCategoryEntity toEntity() {
